Use TiddlyWiki's built in rendering support
The idea is that the main window will contain the config TiddlyWiki. It
will normally be hidden, but can be opened to inspect or update the
configuration.

We’ll add new methods for creating a window corresponding to a given
tiddler (which we’ll use for the wiki listing and the preferences
dialogue) and for creating a window corresponding to a TW file.