Non-functional WIP: experimental code to create a new wiki
Inception: `slave_wiki.js` is an ambitious attempt to run a TW instance
within TW in order to create new wikis.

It doesn’t work at the moment. I suspect that the slave instance is
overwriting some global state that disturbs the operation of the main
backstage wiki.

We may need to explore more thorough ways of sandboxing the wiki
creation (eg shelling to the command line).