Refactoring wiki file support
Now we have a consistent mechanism for wiki file and folder windows.
Next step is to bring backstage tiddler windows into the new structure.

The main user visible change is that we now show the favicon of wiki
file windows, instead of a thumbnail. I’ve found that the thumbnails
were not particularly recognisable, and they were expensive in
performance terms.