Improve window close/quit logic
Previously we were trying to quit when the last window was closed
(excluding the main backstage window). The trouble was that we couldn’t
tell when a wiki folder window was closed. Now instead we quit when the
wiki list window is closed, regardless of the state of other windows.